Delete both saved search AND default custom view upon saved search deletion

+++ src/sentry/api/endpoints/organization_pinned_searches.py
@@ -85,7 +85,5 @@ class OrganizationPinnedSearchEndpoint(OrganizationEndpoint):
             type=search_type.value,
             visibility=Visibility.OWNER_PINNED,
         ).delete()
-        GroupSearchView.objects.filter(
-            organization=organization, user_id=request.user.id, position=0
-        ).delete()
+        GroupSearchView.objects.filter(organization=organization, user_id=request.user.id).delete()
         return Response(status=204)
